geotiff-processor-poc/requirements.txt
rasterio>=1.3.0
Pillow>=9.0.0
numpy>=1.21.0
pyproj>=3.4.0
click>=8.0.0